requirements.txt
selenium==3.141.0
beautifulsoup4==4.9.3
bs4==0.0.1
pandas==1.1.4
sqlalchemy==1.3.20
psycopg2==2.8.6
http-request-randomizer
pymongo
dnspython-2.0.0
PyPDF2-1.26.0
flask
pymysql
python-decouple==3.3
